+++ src/main/java/com/project/sundo_project/SundoProjectApplication.java
@@ -2,8 +2,10 @@ package com.project.sundo_project;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
+@ComponentScan(basePackages = {"com.project.sundo_project", "com.project.sundo_project.config"})
 public class SundoProjectApplication {
 
 	public static void main(String[] args) {
