--- src/main/java/com/project/sundo_project/company/config/CrossOriginConfig.java
@@ -1,23 +1,0 @@
-package com.project.sundo_project.company.config;
-
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-public class CrossOriginConfig implements WebMvcConfigurer {
-    private String[] urls = {
-            "http://localhost:8080"
-    };
-
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-
-        registry
-                .addMapping("/**")
-                .allowedOrigins(urls)
-                .allowedMethods("*")
-                .allowedHeaders("*")
-                .allowCredentials(true)
-        ;
-    }
-
-}
