+++ src/main/java/com/project/sundo_project/evaluation/repository/EvaluationRepository.java
@@ -1,0 +1,7 @@
+package com.project.sundo_project.evaluation.repository;
+
+import com.project.sundo_project.evaluation.entity.Evaluation;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface EvaluationRepository extends JpaRepository<Evaluation,Long> {
+}
