+++ src/main/java/com/project/sundo_project/location/generator/repository/GeneratorRepository.java
@@ -1,0 +1,9 @@
+package com.project.sundo_project.location.generator.repository;
+
+import com.project.sundo_project.location.generator.entity.Generator;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface GeneratorRepository extends JpaRepository<Generator,Long> {
+}
