+++ src/main/java/com/project/sundo_project/location/location/repository/LocationRepository.java
@@ -1,0 +1,10 @@
+package com.project.sundo_project.location.location.repository;
+
+import com.project.sundo_project.location.location.entity.Location;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface LocationRepository extends JpaRepository<Location, Long> {
+
+}
