--- src/main/java/com/project/sundo_project/project/config/SecurityConfig.java
@@ -1,33 +1,0 @@
-package com.project.sundo_project.project.config;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.web.SecurityFilterChain;
-
-@EnableWebSecurity
-@RequiredArgsConstructor
-public class SecurityConfig {
-
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-
-
-        http
-                .cors()
-                .and()
-                .csrf().disable()
-                .httpBasic().disable()
-                .formLogin().disable()
-                .authorizeRequests()
-                .antMatchers("/**").permitAll()
-                .anyRequest().authenticated() // 인가 설정 on
-        ;
-
-
-        return http.build();
-    }
-
-
-}
