--- src/main/java/com/project/sundo_project/repository/EvaluationRepository.java
@@ -1,7 +1,0 @@
-package com.project.sundo_project.repository;
-
-import com.project.sundo_project.entity.Evaluation;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface EvaluationRepository extends JpaRepository<Evaluation,Long> {
-}
